Fit a Temperature Response Curve (TRC) by an index to get a parameter file

This function uses the equation: $$\text{NEE} \sim r_b \cdot \exp \left( E_0 \left( \left( \frac{1}{T_{\text{ref}} - T_0} \right) - \left( \frac{1}{T_{\text{air}} - T_0} \right) \right) \right)$$

The equation requires air temperature (TA) in degrees Celsius, photosynthetically active radiation (PAR) in \(\mu\)mol m-2 s-1, and net ecosystem exchange (NEE) in \(\mu\)mol m-2 s-1.

Usage

TRC_PARMS_01(
  data.frame = NULL,
  iterations = NULL,
  priors.trc = brms::prior("normal(0.5, 0.3)", nlpar = "E0", lb = 0.01, ub = 1) +
    brms::prior("normal(1.0, 0.3)", nlpar = "Rref", lb = 0.01, ub = 1),
  idx.colname = NULL,
  NEE.colname = NULL,
  TA.colname = NULL,
  PAR.colname = NULL
)

Arguments

data.frame

(dataframe) A dataframe that contains net ecosystem exchange (NEE), an index, air temperature (TA), and photosynthetically active radiation (PAR).

iterations

(numeric) The number of iterations to run brms::brm().

priors.trc

(brmsprior dataframe) The priors for brms::brm() to use. Default priors are as follows:

brms::prior("normal(0.5, 0.3)", nlpar = "E0", lb = 0.01, ub = 1) +
brms::prior("normal(1.0, 0.3)", nlpar = "Rref", lb = 0.01, ub = 1)

idx.colname

(character) The name of the column containing the index.

NEE.colname

(character) The name of the column containing NEE.

TA.colname

(character) The name of the column containing air temperature.

PAR.colname

(character) The name of the column containing PAR.

Value

(dataframe) Dataframe of parameter values by the index used to fit them.

Details

Model parameters are fit using the R package brms.

Rhat (Potential Scale Reduction Factor): Indicates how well the different Markov chains in your analysis have converged to the same posterior distribution. Ideally, Rhat should be close to 1 for all parameters. A high Rhat value suggests potential convergence issues and the need to run the chains longer.

Bulk ESS (Effective Sample Size - Bulk): Estimates the effective number of independent samples from the central part of the posterior distribution.

Tail ESS (Effective Sample Size - Tail): Estimates the effective number of independent samples from the tails of the posterior distribution. Important for assessing the reliability of quantile estimates (e.g., 95% confidence intervals).

Key points to remember: Aim for Rhat close to 1 and high values for both Bulk ESS and Tail ESS.

Examples

# Import flux tower data
tower.data <- read.csv(system.file("extdata", "AMF_US-Skr_BASE_HH_2-5_Formatted.csv",
                                   package = "CarbonExchangeParameters"))

# Fit curve parameters for each YearMon:
Example_TRC_PARMS_01 <- TRC_PARMS_01(data.frame = tower.data,
                                     iterations = 1000,
                                     priors.trc = brms::prior("normal(0.5, 0.3)",
                                                    nlpar = "E0", lb = 0.01, ub = 1) +
                                                  brms::prior("normal(1.0, 0.3)",
                                                    nlpar = "Rref", lb = 0.01, ub = 1),
                                     idx.colname = 'YearMon',
                                     NEE.colname = 'NEE_PI',
                                     PAR.colname = 'SW_IN',
                                     TA.colname = 'TA_1_1_1')
#> [1] "GREAT JOB! your dataframe contains idx"
#> [1] "YIPEE! your dataframe contains nee"
#> [1] "Hooray! your dataframe contains PAR"
#> [1] "Hooray! your dataframe contains TA"
#> Your dataframe looks good and you are now ready to start fitting models
#> [1] "2004-07"
#>         idx    nee     TA   PAR
#> 1   2004-07     NA 25.540 0.000
#> 2   2004-07     NA 25.451 0.000
#> 3   2004-07     NA 25.276 0.000
#> 4   2004-07     NA 24.870 0.000
#> 5   2004-07     NA 24.482 0.000
#> 6   2004-07     NA 24.315 0.000
#> 7   2004-07     NA 24.255 0.000
#> 8   2004-07     NA 24.284 0.000
#> 9   2004-07     NA 24.303 0.000
#> 10  2004-07     NA 24.120 0.000
#> 11  2004-07     NA 23.975 0.000
#> 12  2004-07  1.915 24.974 0.000
#> 13  2004-07     NA 24.737 0.000
#> 14  2004-07     NA 24.608 0.000
#> 15  2004-07     NA 24.595 0.000
#> 16  2004-07     NA 24.704 0.000
#> 17  2004-07     NA 24.633 0.000
#> 18  2004-07     NA 24.429 0.000
#> 19  2004-07     NA 24.339 0.000
#> 20  2004-07     NA 24.182 0.000
#> 21  2004-07 10.859 23.887 0.000
#> 22  2004-07     NA 23.666 0.000
#> 23  2004-07     NA 23.522 0.000
#> 24  2004-07     NA 23.553 0.000
#> 25  2004-07     NA 23.417 0.000
#> 26  2004-07     NA 23.567 0.000
#> 27  2004-07     NA 23.583 0.000
#> 28  2004-07     NA 23.523 0.000
#> 29  2004-07     NA 23.613 0.000
#> 30  2004-07     NA 23.496 0.000
#> 31  2004-07     NA 23.401 0.000
#> 32  2004-07     NA 23.370 0.000
#> 33  2004-07     NA 23.319 0.000
#> 34  2004-07     NA 23.181 0.000
#> 35  2004-07     NA 23.075 9.256
#> 36  2004-07     NA 24.701 0.000
#> 37  2004-07     NA 24.350 0.000
#> 38  2004-07     NA 24.574 0.000
#> 39  2004-07     NA 24.621 0.000
#> 40  2004-07     NA 24.396 0.000
#> 41  2004-07     NA 24.274 0.000
#> 42  2004-07     NA 24.573 0.000
#> 43  2004-07     NA 24.372 0.000
#> 44  2004-07     NA 24.161 0.000
#> 45  2004-07     NA 24.118 0.000
#> 46  2004-07     NA 23.718 0.000
#> 47  2004-07     NA 23.480 0.000
#> 48  2004-07     NA 23.344 0.000
#> 49  2004-07     NA 23.303 0.000
#> 50  2004-07     NA 23.367 0.000
#> 51  2004-07     NA 23.339 0.000
#> 52  2004-07     NA 23.283 0.000
#> 53  2004-07  2.425 23.252 0.000
#> 54  2004-07  0.832 23.312 0.000
#> 55  2004-07     NA 23.318 0.000
#> 56  2004-07     NA 23.245 0.000
#> 57  2004-07     NA 23.275 0.000
#> 58  2004-07     NA 23.237 0.000
#> 59  2004-07     NA 23.195 8.006
#> 60  2004-07     NA 27.084 0.000
#> 61  2004-07     NA 26.863 0.000
#> 62  2004-07     NA 26.542 0.000
#> 63  2004-07     NA 26.455 0.000
#> 64  2004-07     NA 26.259 0.000
#> 65  2004-07     NA 25.988 0.000
#> 66  2004-07     NA 25.604 0.000
#> 67  2004-07     NA 25.304 0.000
#> 68  2004-07     NA 24.877 0.000
#> 69  2004-07     NA 24.721 0.000
#> 70  2004-07     NA 24.513 0.000
#> 71  2004-07     NA 24.416 0.000
#> 72  2004-07     NA 24.184 0.000
#> 73  2004-07     NA 24.196 0.000
#> 74  2004-07     NA 24.305 0.000
#> 75  2004-07     NA 24.280 0.000
#> 76  2004-07     NA 24.341 0.000
#> 77  2004-07     NA 24.290 0.000
#> 78  2004-07     NA 24.567 0.000
#> 79  2004-07     NA 24.657 0.000
#> 80  2004-07     NA 24.523 0.000
#> 81  2004-07     NA 24.327 0.000
#> 82  2004-07     NA 24.185 9.116
#> 83  2004-07     NA 25.381 0.000
#> 84  2004-07     NA 25.151 0.000
#> 85  2004-07     NA 25.011 0.000
#> 86  2004-07     NA 24.714 0.000
#> 87  2004-07     NA 24.880 0.000
#> 88  2004-07     NA 24.794 0.000
#> 89  2004-07     NA 24.574 0.000
#> 90  2004-07     NA 24.633 0.000
#> 91  2004-07     NA 24.170 0.000
#> 92  2004-07     NA 24.213 0.000
#> 93  2004-07     NA 24.135 0.000
#> 94  2004-07     NA 23.986 0.000
#> 95  2004-07     NA 24.050 0.000
#> 96  2004-07     NA 24.139 0.000
#> 97  2004-07     NA 24.300 0.000
#> 98  2004-07     NA 24.390 0.000
#> 99  2004-07     NA 24.303 0.000
#> 100 2004-07     NA 24.613 0.000
#> 101 2004-07     NA 25.087 0.000
#> 102 2004-07     NA 25.143 0.000
#> 103 2004-07     NA 24.964 0.000
#> 104 2004-07     NA 24.623 0.000
#> 105 2004-07     NA 24.403 7.551
#> 106 2004-07     NA 26.660 0.000
#> 107 2004-07     NA 26.085 0.000
#> 108 2004-07     NA 25.754 0.000
#> 109 2004-07     NA 25.520 0.000
#> 110 2004-07     NA 25.372 0.000
#> 111 2004-07     NA 25.318 0.000
#> 112 2004-07     NA 25.396 0.000
#> 113 2004-07     NA 25.120 0.000
#> 114 2004-07     NA 24.904 0.000
#> 115 2004-07     NA 24.678 0.000
#> 116 2004-07     NA 24.657 0.000
#> 117 2004-07     NA 24.510 0.000
#> 118 2004-07     NA 24.287 0.000
#> 119 2004-07 10.663 23.941 0.000
#> 120 2004-07     NA 23.751 0.000
#> 121 2004-07     NA 23.873 0.000
#> 122 2004-07     NA 23.817 0.000
#> 123 2004-07     NA 23.690 0.000
#> 124 2004-07     NA 23.747 0.000
#> 125 2004-07     NA 23.756 0.000
#> 126 2004-07     NA 23.816 0.000
#> 127 2004-07     NA 23.790 0.000
#> 128 2004-07     NA 23.662 7.200
#> 129 2004-07  1.129 29.157 0.000
#> 130 2004-07     NA 28.628 0.000
#> 131 2004-07     NA 28.114 0.000
#> 132 2004-07     NA 27.461 0.000
#> 133 2004-07     NA 27.057 0.000
#> 134 2004-07     NA 26.985 0.000
#> 135 2004-07     NA 26.742 0.000
#> 136 2004-07     NA 26.327 0.000
#> 137 2004-07     NA 25.651 0.000
#> 138 2004-07     NA 25.008 0.000
#> 139 2004-07     NA 24.584 0.000
#> 140 2004-07     NA 24.301 0.000
#> 141 2004-07     NA 23.917 0.000
#> 142 2004-07     NA 23.818 0.000
#> 143 2004-07     NA 23.851 0.000
#> 144 2004-07     NA 23.620 0.000
#> 145 2004-07     NA 23.588 0.000
#> 146 2004-07     NA 23.820 0.000
#> 147 2004-07     NA 23.801 0.000
#> 148 2004-07     NA 23.559 0.000
#> 149 2004-07     NA 23.833 0.000
#> 150 2004-07     NA 23.805 0.000
#> 151 2004-07     NA 28.627 0.000
#> 152 2004-07     NA 27.709 0.000
#> 153 2004-07     NA 26.952 0.000
#> 154 2004-07     NA 26.328 0.000
#> 155 2004-07     NA 25.942 0.000
#> 156 2004-07     NA 25.554 0.000
#> 157 2004-07     NA 25.214 0.000
#> 158 2004-07     NA 25.042 0.000
#> 159 2004-07     NA 24.955 0.000
#> 160 2004-07     NA 24.682 0.000
#> 161 2004-07     NA 24.639 0.000
#> 162 2004-07     NA 25.066 0.000
#> 163 2004-07     NA 24.692 0.000
#> 164 2004-07     NA 24.628 0.000
#> 165 2004-07     NA 24.094 0.000
#> 166 2004-07     NA 23.737 0.000
#> 167 2004-07     NA 23.793 0.000
#> 168 2004-07     NA 23.631 0.000
#> 169 2004-07     NA 23.429 0.000
#> 170 2004-07     NA 23.271 0.000
#> 171 2004-07     NA 23.432 0.000
#> 172 2004-07     NA 23.267 0.000
#> 173 2004-07     NA 23.290 6.065
#> 174 2004-07  0.132 30.131 0.000
#> 175 2004-07  1.198 29.444 0.000
#> 176 2004-07  1.342 28.813 0.000
#> 177 2004-07     NA 28.070 0.000
#> 178 2004-07  1.684 27.722 0.000
#> 179 2004-07     NA 27.323 0.000
#> 180 2004-07     NA 26.680 0.000
#> 181 2004-07     NA 25.828 0.000
#> 182 2004-07     NA 25.344 0.000
#> 183 2004-07     NA 25.466 0.000
#> 184 2004-07     NA 25.467 0.000
#> 185 2004-07  1.706 25.636 0.000
#> 186 2004-07     NA 25.651 0.000
#> 187 2004-07     NA 25.456 0.000
#> 188 2004-07     NA 25.256 0.000
#> 189 2004-07     NA 25.156 0.000
#> 190 2004-07     NA 24.991 0.000
#> 191 2004-07     NA 24.945 0.000
#> 192 2004-07     NA 24.991 0.000
#> 193 2004-07     NA 24.773 0.000
#> 194 2004-07     NA 24.764 0.000
#> 195 2004-07     NA 24.768 0.000
#> 196 2004-07     NA 24.536 3.804
#> 197 2004-07     NA 30.372 0.000
#> 198 2004-07     NA 29.237 0.000
#> 199 2004-07     NA 28.461 0.000
#> 200 2004-07     NA 27.851 0.000
#> 201 2004-07  4.593 27.389 0.000
#> 202 2004-07  3.137 27.106 0.000
#> 203 2004-07     NA 26.928 0.000
#> 204 2004-07     NA 26.686 0.000
#> 205 2004-07     NA 26.236 0.000
#> 206 2004-07     NA 25.954 0.000
#> 207 2004-07     NA 25.732 0.000
#> 208 2004-07     NA 25.441 0.000
#> 209 2004-07     NA 25.248 0.000
#> 210 2004-07     NA 25.204 0.000
#> 211 2004-07     NA 25.128 0.000
#> 212 2004-07     NA 25.098 0.000
#> 213 2004-07     NA 25.219 0.000
#> 214 2004-07     NA 25.269 0.000
#> 215 2004-07     NA 25.216 0.000
#> 216 2004-07     NA 25.195 0.000
#> 217 2004-07     NA 25.123 0.000
#> 218 2004-07     NA 25.100 0.000
#> 219 2004-07     NA 25.165 6.896
#> 220 2004-07     NA 23.723 0.000
#> 221 2004-07     NA 23.623 0.000
#> 222 2004-07     NA 23.670 0.000
#> 223 2004-07     NA 23.948 0.000
#> 224 2004-07     NA 23.790 0.000
#> 225 2004-07     NA 23.490 0.000
#> 226 2004-07     NA 23.280 0.000
#> 227 2004-07     NA 23.050 0.000
#> 228 2004-07     NA 23.016 0.000
#> 229 2004-07     NA 23.151 0.000
#> 230 2004-07     NA 23.232 0.000
#> 231 2004-07     NA 22.996 0.000
#> 232 2004-07     NA 22.886 0.000
#> 233 2004-07  8.524 22.702 0.000
#> 234 2004-07  9.591 22.707 0.000
#> 235 2004-07     NA 22.853 0.000
#> 236 2004-07     NA 22.881 0.000
#> 237 2004-07  8.993 22.655 0.000
#> 238 2004-07     NA 22.610 0.000
#> 239 2004-07     NA 22.731 0.000
#> 240 2004-07     NA 22.799 0.000
#> 241 2004-07     NA 22.699 0.000
#> 242 2004-07     NA 22.592 5.126
#> 243 2004-07     NA 21.487 6.538
#> 244 2004-07     NA 22.264 4.797
#> 245 2004-07     NA 22.088 5.883
#> 246 2004-07     NA 23.585 0.000
#> 247 2004-07     NA 23.334 0.000
#> 248 2004-07     NA 23.132 0.000
#> 249 2004-07     NA 23.043 0.000
#> 250 2004-07     NA 23.073 0.000
#> 251 2004-07     NA 23.131 0.000
#> 252 2004-07     NA 23.018 0.000
#> 253 2004-07     NA 22.911 0.000
#> 254 2004-07     NA 22.655 0.000
#> 255 2004-07     NA 22.432 0.000
#> 256 2004-07     NA 22.230 0.000
#> 257 2004-07     NA 22.227 0.000
#> 258 2004-07     NA 22.195 0.000
#> 259 2004-07     NA 22.041 0.000
#> 260 2004-07     NA 21.921 0.000
#> 261 2004-07     NA 21.953 0.000
#> 262 2004-07     NA 21.990 0.000
#> 263 2004-07     NA 21.869 0.000
#> 264 2004-07     NA 21.817 0.000
#> 265 2004-07     NA 21.780 0.000
#> 266 2004-07     NA 21.613 0.000
#> 267 2004-07     NA 21.417 0.000
#> 268 2004-07     NA 21.612 4.694
#> 269 2004-07     NA 28.319 0.000
#> 270 2004-07     NA 27.443 0.000
#> 271 2004-07     NA 26.724 0.000
#> 272 2004-07     NA 26.582 0.000
#> 273 2004-07     NA 26.146 0.000
#> 274 2004-07     NA 25.672 0.000
#> 275 2004-07     NA 25.925 0.000
#> 276 2004-07     NA 26.034 0.000
#> 277 2004-07     NA 25.434 0.000
#> 278 2004-07     NA 25.333 0.000
#> 279 2004-07     NA 25.069 0.000
#> 280 2004-07     NA 24.851 0.000
#> 281 2004-07     NA 24.513 0.000
#> 282 2004-07     NA 24.458 0.000
#> 283 2004-07     NA 24.221 0.000
#> 284 2004-07     NA 24.322 0.000
#> 285 2004-07     NA 24.575 0.000
#> 286 2004-07     NA 24.362 0.000
#> 287 2004-07     NA 23.858 0.000
#> 288 2004-07     NA 23.496 0.000
#> 289 2004-07     NA 23.355 0.000
#> 290 2004-07     NA 23.255 0.000
#> 291 2004-07     NA 23.331 1.939
#> 292 2004-07     NA 28.586 0.000
#> 293 2004-07     NA 28.184 0.000
#> 294 2004-07     NA 27.464 0.000
#> 295 2004-07     NA 26.852 0.000
#> 296 2004-07     NA 26.577 0.000
#> 297 2004-07     NA 26.491 0.000
#> 298 2004-07     NA 26.460 0.000
#> 299 2004-07     NA 26.105 0.000
#> 300 2004-07     NA 26.645 0.000
#> 301 2004-07     NA 26.561 0.000
#> 302 2004-07     NA 26.350 0.000
#> 303 2004-07     NA 26.271 0.000
#> 304 2004-07     NA 25.732 0.000
#> 305 2004-07     NA 25.328 0.000
#> 306 2004-07     NA 25.363 0.000
#> 307 2004-07     NA 25.247 0.000
#> 308 2004-07     NA 25.181 0.000
#> 309 2004-07     NA 25.098 0.000
#> 310 2004-07     NA 24.964 0.000
#> 311 2004-07     NA 24.900 0.000
#> 312 2004-07     NA 24.870 0.000
#> 313 2004-07     NA 24.880 0.000
#> 314 2004-07     NA 24.735 2.590
#> 315 2004-07  1.092 29.701 0.000
#> 316 2004-07  3.176 28.990 0.000
#> 317 2004-07  3.292 28.403 0.000
#> 318 2004-07     NA 27.972 0.000
#> 319 2004-07     NA 27.639 0.000
#> 320 2004-07     NA 26.966 0.000
#> 321 2004-07     NA 27.353 0.000
#> 322 2004-07     NA 26.682 0.000
#> 323 2004-07     NA 27.239 0.000
#> 324 2004-07     NA 26.812 0.000
#> 325 2004-07     NA 26.618 0.000
#> 326 2004-07     NA 26.342 0.000
#> 327 2004-07     NA 26.651 0.000
#> 328 2004-07     NA 26.397 0.000
#> 329 2004-07     NA 26.434 0.000
#> 330 2004-07     NA 26.261 0.000
#> 331 2004-07     NA 26.233 0.000
#> 332 2004-07     NA 26.124 0.000
#> 333 2004-07     NA 25.673 0.000
#> 334 2004-07     NA 25.930 0.000
#> 335 2004-07     NA 26.242 0.000
#> 336 2004-07     NA 25.431 0.000
#> 337 2004-07     NA 25.233 3.624
#> 338 2004-07     NA 28.750 0.000
#> 339 2004-07     NA 28.776 0.000
#> 340 2004-07     NA 28.451 0.000
#> 341 2004-07     NA 28.190 0.000
#> 342 2004-07     NA 27.950 0.000
#> 343 2004-07     NA 27.830 0.000
#> 344 2004-07     NA 27.100 0.000
#> 345 2004-07     NA 26.757 0.000
#> 346 2004-07     NA 26.407 0.000
#> 347 2004-07     NA 26.161 0.000
#> 348 2004-07     NA 26.519 0.000
#> 349 2004-07  0.711 26.809 0.000
#> 350 2004-07     NA 26.218 0.000
#> 351 2004-07     NA 26.032 0.000
#> 352 2004-07     NA 25.958 0.000
#> 353 2004-07     NA 25.968 0.000
#> 354 2004-07     NA 25.574 0.000
#> 355 2004-07     NA 25.404 0.000
#> 356 2004-07     NA 25.498 0.000
#> 357 2004-07     NA 25.863 0.000
#> 358 2004-07     NA 25.576 0.000
#> 359 2004-07     NA 25.563 0.000
#> 360 2004-07     NA 25.553 0.000
#> 361 2004-07     NA 25.481 3.506
#> 362 2004-07     NA 29.806 0.000
#> 363 2004-07     NA 28.393 0.000
#> 364 2004-07     NA 26.729 0.000
#> 365 2004-07     NA 25.822 0.000
#> 366 2004-07     NA 25.443 0.000
#> 367 2004-07     NA 25.232 0.000
#> 368 2004-07  3.487 24.960 0.000
#> 369 2004-07     NA 24.731 0.000
#> 370 2004-07     NA 24.619 0.000
#> 371 2004-07     NA 24.493 0.000
#> 372 2004-07     NA 24.584 0.000
#> 373 2004-07     NA 24.474 0.000
#> 374 2004-07     NA 24.419 0.000
#> 375 2004-07     NA 24.301 0.000
#> 376 2004-07  3.713 24.367 0.000
#> 377 2004-07     NA 24.312 0.000
#> 378 2004-07     NA 24.372 0.000
#> 379 2004-07     NA 24.728 0.000
#> 380 2004-07     NA 24.846 0.000
#> 381 2004-07     NA 24.817 0.000
#> 382 2004-07     NA 24.939 0.000
#> 383 2004-07     NA 24.880 0.000
#> 384 2004-07     NA 25.337 0.000
#> 385 2004-07     NA 25.120 2.989
#> 386 2004-07 -0.802 29.595 0.000
#> 387 2004-07  2.905 29.004 0.000
#> 388 2004-07  2.945 28.743 0.000
#> 389 2004-07     NA 27.538 0.000
#> 390 2004-07     NA 25.055 0.000
#> 391 2004-07     NA 25.208 0.000
#> 392 2004-07     NA 25.714 0.000
#> 393 2004-07     NA 25.801 0.000
#> 394 2004-07     NA 25.389 0.000
#> 395 2004-07     NA 25.428 0.000
#> 396 2004-07     NA 26.022 0.000
#> 397 2004-07     NA 26.162 0.000
#> 398 2004-07     NA 26.459 0.000
#> 399 2004-07  6.515 26.490 0.000
#> 400 2004-07 -1.965 25.694 0.000
#> 401 2004-07     NA 25.912 0.000
#> 402 2004-07     NA 26.183 0.000
#> 403 2004-07  1.854 26.198 0.000
#> 404 2004-07 14.794 26.211 0.000
#> 405 2004-07 10.912 25.829 0.000
#> 406 2004-07 12.455 25.632 0.000
#> 407 2004-07     NA 25.120 0.000
#> 408 2004-07     NA 23.944 7.338
#> 409 2004-07     NA 23.183 9.658
#> 410 2004-07     NA 23.680 9.572
#> 411 2004-07     NA 23.182 0.000
#> 412 2004-07     NA 23.437 0.000
#> 413 2004-07     NA 23.809 0.000
#> 414 2004-07     NA 24.409 0.000
#> 415 2004-07     NA 24.075 0.000
#> 416 2004-07     NA 23.788 0.000
#> 417 2004-07     NA 23.737 0.000
#> 418 2004-07     NA 23.685 0.000
#> 419 2004-07     NA 24.145 0.000
#> 420 2004-07     NA 24.508 0.000
#> 421 2004-07     NA 24.690 0.000
#> 422 2004-07     NA 25.007 0.000
#> 423 2004-07     NA 25.257 0.000
#> 424 2004-07     NA 25.340 0.000
#> 425 2004-07     NA 25.326 0.000
#> 426 2004-07     NA 25.378 0.000
#> 427 2004-07     NA 25.713 0.000
#> 428 2004-07     NA 25.423 0.000
#> 429 2004-07     NA 25.653 0.000
#> 430 2004-07     NA 25.716 0.000
#> 431 2004-07     NA 25.126 0.000
#> 432 2004-07     NA 24.861 1.913
#> 433 2004-07     NA 24.126 4.747
#> 434 2004-07 -1.817 29.009 0.000
#> 435 2004-07     NA 28.552 0.000
#> 436 2004-07     NA 26.606 0.000
#> 437 2004-07     NA 23.968 0.000
#> 438 2004-07     NA 23.664 0.000
#> 439 2004-07     NA 23.754 0.000
#> 440 2004-07     NA 23.841 0.000
#> 441 2004-07     NA 23.763 0.000
#> 442 2004-07     NA 23.889 0.000
#> 443 2004-07     NA 24.165 0.000
#> 444 2004-07     NA 24.589 0.000
#> 445 2004-07     NA 24.830 0.000
#> 446 2004-07     NA 24.567 0.000
#> 447 2004-07     NA 24.253 0.000
#> 448 2004-07     NA 24.032 0.000
#> 449 2004-07     NA 23.882 0.000
#> 450 2004-07     NA 23.720 0.000
#> 451 2004-07     NA 24.001 0.000
#> 452 2004-07     NA 24.341 0.000
#> 453 2004-07     NA 24.176 0.000
#> 454 2004-07     NA 24.020 0.000
#> 455 2004-07     NA 23.976 0.000
#> 456 2004-07     NA 23.753 0.000
#> 457 2004-07     NA 23.842 0.184
#> 458 2004-07     NA 23.716 1.483
#> 459 2004-07     NA 23.002 1.959
#> 460 2004-07     NA 22.983 6.546
#> 461 2004-07     NA 27.533 0.000
#> 462 2004-07  6.402 26.913 0.000
#> 463 2004-07     NA 26.102 0.000
#> 464 2004-07     NA 25.762 0.000
#> 465 2004-07     NA 25.246 0.000
#> 466 2004-07     NA 25.164 0.000
#> 467 2004-07     NA 25.181 0.000
#> 468 2004-07     NA 25.182 0.000
#> 469 2004-07     NA 25.000 0.000
#> 470 2004-07     NA 24.463 0.000
#> 471 2004-07     NA 23.139 0.000
#> 472 2004-07     NA 22.992 0.000
#> 473 2004-07     NA 22.879 0.000
#> 474 2004-07     NA 22.801 0.000
#> 475 2004-07     NA 22.744 0.000
#> 476 2004-07     NA 22.693 0.000
#> 477 2004-07     NA 22.658 0.000
#> 478 2004-07     NA 22.642 0.000
#> 479 2004-07     NA 22.610 0.000
#> 480 2004-07     NA 22.520 0.000
#> 481 2004-07     NA 22.350 0.000
#> 482 2004-07     NA 22.114 0.000
#> 483 2004-07     NA 21.955 0.000
#> 484 2004-07     NA 22.032 2.499
#> 485 2004-07     NA 23.797 0.000
#> 486 2004-07     NA 23.794 0.000
#> 487 2004-07     NA 23.719 0.000
#> 488 2004-07     NA 23.621 0.000
#> 489 2004-07     NA 23.550 0.000
#> 490 2004-07     NA 23.447 0.000
#> 491 2004-07     NA 23.254 0.000
#> 492 2004-07     NA 23.052 0.000
#> 493 2004-07     NA 22.930 0.000
#> 494 2004-07     NA 25.458 0.000
#> 495 2004-07     NA 25.293 0.000
#> 496 2004-07     NA 25.071 0.000
#> 497 2004-07     NA 24.911 0.000
#> 498 2004-07     NA 24.652 0.000
#> 499 2004-07     NA 24.549 0.000
#> 500 2004-07     NA 24.391 0.000
#> 501 2004-07     NA 24.277 0.000
#> 502 2004-07     NA 24.193 0.000
#> 503 2004-07     NA 24.370 0.000
#> 504 2004-07     NA 24.212 0.000
#> 505 2004-07     NA 24.301 0.000
#> 506 2004-07     NA 24.379 0.000
#> 507 2004-07     NA 24.442 3.097
#> 508 2004-07  1.682 27.178 0.000
#> 509 2004-07  4.115 26.792 0.000
#> 510 2004-07  3.676 26.602 0.000
#> 511 2004-07  3.854 26.683 0.000
#> 512 2004-07     NA 26.643 0.000
#> 513 2004-07     NA 26.432 0.000
#> 514 2004-07     NA 26.097 0.000
#> 515 2004-07     NA 25.863 0.000
#> 516 2004-07     NA 25.410 0.000
#> 517 2004-07     NA 25.103 0.000
#> 518 2004-07     NA 24.899 0.000
#> 519 2004-07     NA 24.647 0.000
#> 520 2004-07     NA 24.563 0.000
#> 521 2004-07     NA 24.539 0.000
#> 522 2004-07     NA 24.356 0.000
#> 523 2004-07     NA 24.386 0.000
#> 524 2004-07     NA 24.276 0.000
#> 525 2004-07     NA 24.177 0.000
#> 526 2004-07     NA 24.133 0.000
#> 527 2004-07     NA 23.882 0.000
#> 528 2004-07     NA 23.746 0.000
#> 529 2004-07     NA 23.745 0.000
#> 530 2004-07     NA 23.844 0.000
#> 531 2004-07     NA 24.031 3.091
#> 532 2004-07     NA 28.704 0.000
#> 533 2004-07  2.686 28.359 0.000
#> 534 2004-07  2.035 27.932 0.000
#> 535 2004-07     NA 27.544 0.000
#> 536 2004-07  3.118 27.094 0.000
#> 537 2004-07     NA 26.709 0.000
#> 538 2004-07     NA 26.314 0.000
#> 539 2004-07     NA 26.066 0.000
#> 540 2004-07     NA 25.803 0.000
#> 541 2004-07     NA 25.562 0.000
#> 542 2004-07     NA 25.340 0.000
#> 543 2004-07     NA 25.069 0.000
#> 544 2004-07     NA 24.926 0.000
#> 545 2004-07     NA 24.839 0.000
#> 546 2004-07     NA 24.866 0.000
#> 547 2004-07     NA 24.458 0.000
#> 548 2004-07     NA 24.152 0.000
#> 549 2004-07     NA 24.107 0.000
#> 550 2004-07     NA 24.217 0.000
#> 551 2004-07     NA 24.170 0.000
#> 552 2004-07     NA 23.893 0.000
#> 553 2004-07     NA 23.866 0.000
#> 554 2004-07     NA 23.875 1.586
#> 555 2004-07     NA 23.438 0.000
#> 556 2004-07     NA 23.251 0.000
#> 557 2004-07     NA 23.533 0.000
#> 558 2004-07     NA 23.631 0.000
#> 559 2004-07     NA 23.801 0.000
#> 560 2004-07     NA 23.964 0.000
#> 561 2004-07     NA 24.284 0.000
#> 562 2004-07     NA 24.290 0.000
#> 563 2004-07     NA 24.088 0.000
#> 564 2004-07     NA 24.074 0.000
#> 565 2004-07     NA 24.075 0.000
#> 566 2004-07     NA 24.028 0.000
#> 567 2004-07     NA 24.178 0.000
#> 568 2004-07     NA 24.216 0.000
#> 569 2004-07     NA 24.095 0.000
#> 570 2004-07     NA 24.053 0.000
#> 571 2004-07     NA 24.011 0.000
#> 572 2004-07     NA 23.855 0.000
#> 573 2004-07     NA 23.745 0.000
#> 574 2004-07     NA 23.727 0.000
#> 575 2004-07     NA 23.656 0.000
#> 576 2004-07     NA 23.807 0.000
#> 577 2004-07     NA 23.923 0.000
#> 578 2004-07     NA 27.051 0.000
#> 579 2004-07     NA 26.788 0.000
#> 580 2004-07     NA 26.203 0.000
#> 581 2004-07     NA 25.552 0.000
#> 582 2004-07     NA 25.295 0.000
#> 583 2004-07     NA 25.371 0.000
#> 584 2004-07     NA 25.410 0.000
#> 585 2004-07     NA 25.024 0.000
#> 586 2004-07     NA 24.552 0.000
#> 587 2004-07     NA 24.368 0.000
#> 588 2004-07     NA 24.255 0.000
#> 589 2004-07     NA 24.239 0.000
#> 590 2004-07     NA 24.389 0.000
#> 591 2004-07     NA 24.303 0.000
#> 592 2004-07     NA 23.967 0.000
#> 593 2004-07     NA 24.063 0.000
#> 594 2004-07     NA 24.293 0.000
#> 595 2004-07     NA 24.485 0.000
#> 596 2004-07     NA 24.391 0.000
#> 597 2004-07     NA 24.044 0.000
#> 598 2004-07     NA 23.829 0.000
#> 599 2004-07     NA 23.418 0.000
#> 600 2004-07     NA 23.207 0.000
#> 601 2004-07     NA 23.371 2.477
#> 602 2004-07     NA 26.901 9.745
#> 603 2004-07     NA 24.599 5.686
#> 604 2004-07  7.263 23.676 8.519
#> 605 2004-07     NA 23.451 0.000
#> 606 2004-07     NA 23.479 0.000
#> 607 2004-07     NA 23.286 0.000
#> 608 2004-07     NA 23.212 0.000
#> 609 2004-07     NA 23.064 0.000
#> 610 2004-07     NA 23.051 0.000
#> 611 2004-07     NA 23.129 0.000
#> 612 2004-07     NA 23.156 0.000
#> 613 2004-07     NA 23.415 0.000
#> 614 2004-07     NA 23.442 0.000
#> 615 2004-07     NA 23.101 0.000
#> 616 2004-07     NA 23.014 0.000
#> 617 2004-07     NA 23.189 0.000
#> 618 2004-07     NA 23.156 0.000
#> 619 2004-07     NA 23.329 0.000
#> 620 2004-07     NA 23.306 0.000
#> 621 2004-07     NA 23.289 0.000
#> 622 2004-07     NA 23.276 0.000
#> 623 2004-07     NA 23.128 0.000
#> 624 2004-07     NA 23.162 0.000
#> 625 2004-07     NA 23.192 0.000
#> 626 2004-07     NA 23.243 0.000
#> 627 2004-07     NA 23.271 0.000
#> 628 2004-07     NA 23.360 0.591
#> 629 2004-07     NA 27.154 0.000
#> 630 2004-07     NA 26.015 0.000
#> 631 2004-07     NA 25.835 0.000
#> 632 2004-07     NA 25.657 0.000
#> 633 2004-07     NA 25.431 0.000
#> 634 2004-07     NA 24.487 0.000
#> 635 2004-07     NA 24.173 0.000
#> 636 2004-07     NA 24.060 0.000
#> 637 2004-07     NA 24.044 0.000
#> 638 2004-07     NA 24.113 0.000
#> 639 2004-07     NA 23.962 0.000
#> 640 2004-07     NA 23.758 0.000
#> 641 2004-07     NA 23.632 0.000
#> 642 2004-07     NA 23.667 0.000
#> 643 2004-07     NA 23.780 0.000
#> 644 2004-07     NA 23.718 0.000
#> 645 2004-07     NA 23.695 0.000
#> 646 2004-07     NA 23.656 0.000
#> 647 2004-07     NA 23.649 0.000
#> 648 2004-07     NA 24.033 0.000
#> 649 2004-07     NA 24.055 0.000
#> 650 2004-07     NA 23.794 0.000
#> 651 2004-07     NA 23.573 0.000
#> 652 2004-07     NA 23.509 0.499
#> 653 2004-07     NA 26.163 0.000
#> 654 2004-07     NA 25.947 0.000
#> 655 2004-07     NA 25.537 0.000
#> 656 2004-07     NA 25.335 0.000
#> 657 2004-07     NA 25.119 0.000
#> 658 2004-07     NA 24.699 0.000
#> 659 2004-07     NA 24.899 0.000
#> 660 2004-07     NA 24.715 0.000
#> 661 2004-07     NA 24.904 0.000
#> 662 2004-07     NA 24.755 0.000
#> 663 2004-07     NA 24.679 0.000
#> 664 2004-07     NA 24.596 0.000
#> 665 2004-07     NA 24.724 0.000
#> 666 2004-07     NA 24.664 0.000
#> 667 2004-07     NA 24.708 0.000
#> 668 2004-07     NA 24.527 0.000
#> 669 2004-07     NA 24.511 0.000
#> 670 2004-07     NA 24.293 0.000
#> 671 2004-07     NA 24.028 0.000
#> 672 2004-07     NA 23.948 0.000
#> 673 2004-07     NA 23.954 0.000
#> 674 2004-07     NA 23.859 0.000
#> 675 2004-07     NA 23.753 0.000
#> 676 2004-07     NA 23.553 0.754
#> 677 2004-07     NA 27.002 0.000
#> 678 2004-07  1.259 26.657 0.000
#> 679 2004-07  2.217 26.447 0.000
#> 680 2004-07     NA 26.221 0.000
#> 681 2004-07     NA 26.014 0.000
#> 682 2004-07     NA 25.692 0.000
#> 683 2004-07     NA 25.612 0.000
#> 684 2004-07     NA 25.530 0.000
#> 685 2004-07     NA 25.420 0.000
#> 686 2004-07     NA 25.320 0.000
#> 687 2004-07     NA 25.220 0.000
#> 688 2004-07     NA 24.943 0.000
#> 689 2004-07     NA 24.647 0.000
#> 690 2004-07     NA 24.521 0.000
#> 691 2004-07     NA 24.488 0.000
#> 692 2004-07     NA 24.585 0.000
#> 693 2004-07     NA 24.493 0.000
#> 694 2004-07     NA 24.453 0.000
#> 695 2004-07     NA 24.250 0.000
#> 696 2004-07     NA 24.167 0.000
#> 697 2004-07     NA 23.997 0.000
#> 698 2004-07     NA 23.863 0.000
#> 699 2004-07     NA 23.735 0.000
#> 700 2004-07     NA 23.770 0.944
#> 701 2004-07     NA 23.926 9.373
#> 702 2004-07  3.315 28.161 0.000
#> 703 2004-07  1.571 27.866 0.000
#> 704 2004-07  3.981 27.670 0.000
#> 705 2004-07  3.346 27.486 0.000
#> 706 2004-07     NA 27.273 0.000
#> 707 2004-07     NA 26.858 0.000
#> 708 2004-07     NA 26.557 0.000
#> 709 2004-07     NA 25.915 0.000
#> 710 2004-07     NA 25.766 0.000
#> 711 2004-07     NA 25.560 0.000
#> 712 2004-07  5.595 25.388 0.000
#> 713 2004-07     NA 25.064 0.000
#> 714 2004-07     NA 24.859 0.000
#> 715 2004-07     NA 24.895 0.000
#> 716 2004-07     NA 24.675 0.000
#> 717 2004-07     NA 24.819 0.000
#> 718 2004-07     NA 24.753 0.000
#> 719 2004-07     NA 24.618 0.000
#> 720 2004-07     NA 24.522 0.000
#> 721 2004-07     NA 24.507 0.000
#> 722 2004-07     NA 24.754 0.000
#> 723 2004-07     NA 25.373 0.000
#> 724 2004-07     NA 25.555 0.313
#> 725 2004-07     NA 26.307 0.000
#> 726 2004-07     NA 26.092 0.000
#> 727 2004-07 11.472 25.429 0.000
#> 728 2004-07  7.672 23.551 0.000
#> 729 2004-07     NA 23.446 0.000
#> 730 2004-07     NA 23.554 0.000
#> 731 2004-07     NA 24.405 0.000
#> 732 2004-07     NA 24.240 0.000
#> 733 2004-07     NA 24.387 0.000
#> 734 2004-07     NA 24.496 0.000
#> 735 2004-07     NA 24.413 0.000
#> 736 2004-07     NA 24.316 0.000
#> Warning: Rows containing NAs were excluded from the model.
#> Error : CmdStan path has not been set yet. See ?set_cmdstan_path.
#> Error in TRC_PARMS_01(data.frame = tower.data, iterations = 1000, priors.trc = brms::prior("normal(0.5, 0.3)",     nlpar = "E0", lb = 0.01, ub = 1) + brms::prior("normal(1.0, 0.3)",     nlpar = "Rref", lb = 0.01, ub = 1), idx.colname = "YearMon",     NEE.colname = "NEE_PI", PAR.colname = "SW_IN", TA.colname = "TA_1_1_1"): object 'model.brms' not found