Fit a Temperature Response Curve (TRC) by an index to get a parameter file

This function uses the equation: $$\text{NEE} \sim R_{\text{ref}} \cdot \exp \left( \frac{E_a}{R} \left( \frac{1}{T_{\text{ref}}} - \frac{1}{T} \right) \right) $$

Where \(R_{ref}\) is the reference respiration rate [\(\mu\)mol CO2 m-2 s-1], \(E_a\) is the activation energy [J/mol], \(R\) is the Universal Gas Constant [8.314 J mol-1 K-1], \(T_{ref}\) is the reference temperature (25 C or 298K) [K], and TA is temperature [K].

WARNING: TA must be in kelvin.

The equation requires air temperature (TA) in kelvin, photosynthetically active radiation (PAR) in \(\mu\)mol m-2 s-1, and net ecosystem exchange (NEE) in \(\mu\)mol m-2 s-1.

Usage

TRC_PARMS_06(
  data.frame = NULL,
  iterations = NULL,
  priors.trc = brms::prior("normal(0.5, 0.3)", nlpar = "Ea", lb = 0.01, ub = 1) +
    brms::prior("normal(.5, .3)", nlpar = "Rref", lb = 0.01, ub = 1),
  idx.colname = NULL,
  NEE.colname = NULL,
  PAR.colname = NULL,
  TA.colname = NULL
)

Arguments

data.frame

(dataframe) A dataframe that contains net ecosystem exchange (NEE), an index, air temperature (TA), and photosynthetically active radiation (PAR).

iterations

(numeric) The number of iterations to run brms::brm().

priors.trc

(brmsprior dataframe) The priors for brms::brm() to use. Default priors are as follows:

brms::prior("normal(0.5, 0.3)", nlpar = "Ea", lb = 0.01, ub = 1) +
brms::prior("normal(.5, .3)", nlpar = "Rref", lb = 0.01, ub = 1)

idx.colname

(character) The name of the column containing the index.

NEE.colname

(character) The name of the column containing NEE.

PAR.colname

(character) The name of the column containing PAR.

TA.colname

(character) The name of the column containing air temperature.

Value

(dataframe) Dataframe of parameter values by the index used to fit them.

Details

Model parameters are fit using the R package brms.

Rhat (Potential Scale Reduction Factor): Indicates how well the different Markov chains in your analysis have converged to the same posterior distribution. Ideally, Rhat should be close to 1 for all parameters. A high Rhat value suggests potential convergence issues and the need to run the chains longer.

Bulk ESS (Effective Sample Size - Bulk): Estimates the effective number of independent samples from the central part of the posterior distribution.

Tail ESS (Effective Sample Size - Tail): Estimates the effective number of independent samples from the tails of the posterior distribution. Important for assessing the reliability of quantile estimates (e.g., 95% confidence intervals).

Key points to remember: Aim for Rhat close to 1 and high values for both Bulk ESS and Tail ESS.

Examples

# Import flux tower data
tower.data <- read.csv(system.file("extdata", "AMF_US-Skr_BASE_HH_2-5_Formatted.csv",
                                   package = "CarbonExchangeParameters"))

# Fit curve parameters for each YearMon:
Example_TRC_PARMS_06 <- TRC_PARMS_06(data.frame = tower.data,
                                     iterations = 5000,
                                     priors.trc = brms::prior("normal(0.5, 0.3)",
                                                    nlpar = "Ea", lb = 0.01, ub = 1) +
                                                  brms::prior("normal(.5, .3)",
                                                    nlpar = "Rref", lb = 0.01, ub = 1),
                                     idx.colname = 'YearMon',
                                     NEE.colname = 'NEE_PI',
                                     PAR.colname = 'SW_IN',
                                     TA.colname = 'TA_1_1_1')
#> [1] "GREAT JOB! your dataframe contains idx"
#> [1] "YIPEE! your dataframe contains nee"
#> [1] "Hooray! your dataframe contains PAR"
#> [1] "Hooray! your dataframe contains TA, this is a test"
#>  Your dataframe looks good and you are now ready to start fitting models
#> [1] "2004-07"
#>         idx    nee   PAR     TA
#> 1   2004-07     NA 0.000 25.540
#> 2   2004-07     NA 0.000 25.451
#> 3   2004-07     NA 0.000 25.276
#> 4   2004-07     NA 0.000 24.870
#> 5   2004-07     NA 0.000 24.482
#> 6   2004-07     NA 0.000 24.315
#> 7   2004-07     NA 0.000 24.255
#> 8   2004-07     NA 0.000 24.284
#> 9   2004-07     NA 0.000 24.303
#> 10  2004-07     NA 0.000 24.120
#> 11  2004-07     NA 0.000 23.975
#> 12  2004-07  1.915 0.000 24.974
#> 13  2004-07     NA 0.000 24.737
#> 14  2004-07     NA 0.000 24.608
#> 15  2004-07     NA 0.000 24.595
#> 16  2004-07     NA 0.000 24.704
#> 17  2004-07     NA 0.000 24.633
#> 18  2004-07     NA 0.000 24.429
#> 19  2004-07     NA 0.000 24.339
#> 20  2004-07     NA 0.000 24.182
#> 21  2004-07 10.859 0.000 23.887
#> 22  2004-07     NA 0.000 23.666
#> 23  2004-07     NA 0.000 23.522
#> 24  2004-07     NA 0.000 23.553
#> 25  2004-07     NA 0.000 23.417
#> 26  2004-07     NA 0.000 23.567
#> 27  2004-07     NA 0.000 23.583
#> 28  2004-07     NA 0.000 23.523
#> 29  2004-07     NA 0.000 23.613
#> 30  2004-07     NA 0.000 23.496
#> 31  2004-07     NA 0.000 23.401
#> 32  2004-07     NA 0.000 23.370
#> 33  2004-07     NA 0.000 23.319
#> 34  2004-07     NA 0.000 23.181
#> 35  2004-07     NA 9.256 23.075
#> 36  2004-07     NA 0.000 24.701
#> 37  2004-07     NA 0.000 24.350
#> 38  2004-07     NA 0.000 24.574
#> 39  2004-07     NA 0.000 24.621
#> 40  2004-07     NA 0.000 24.396
#> 41  2004-07     NA 0.000 24.274
#> 42  2004-07     NA 0.000 24.573
#> 43  2004-07     NA 0.000 24.372
#> 44  2004-07     NA 0.000 24.161
#> 45  2004-07     NA 0.000 24.118
#> 46  2004-07     NA 0.000 23.718
#> 47  2004-07     NA 0.000 23.480
#> 48  2004-07     NA 0.000 23.344
#> 49  2004-07     NA 0.000 23.303
#> 50  2004-07     NA 0.000 23.367
#> 51  2004-07     NA 0.000 23.339
#> 52  2004-07     NA 0.000 23.283
#> 53  2004-07  2.425 0.000 23.252
#> 54  2004-07  0.832 0.000 23.312
#> 55  2004-07     NA 0.000 23.318
#> 56  2004-07     NA 0.000 23.245
#> 57  2004-07     NA 0.000 23.275
#> 58  2004-07     NA 0.000 23.237
#> 59  2004-07     NA 8.006 23.195
#> 60  2004-07     NA 0.000 27.084
#> 61  2004-07     NA 0.000 26.863
#> 62  2004-07     NA 0.000 26.542
#> 63  2004-07     NA 0.000 26.455
#> 64  2004-07     NA 0.000 26.259
#> 65  2004-07     NA 0.000 25.988
#> 66  2004-07     NA 0.000 25.604
#> 67  2004-07     NA 0.000 25.304
#> 68  2004-07     NA 0.000 24.877
#> 69  2004-07     NA 0.000 24.721
#> 70  2004-07     NA 0.000 24.513
#> 71  2004-07     NA 0.000 24.416
#> 72  2004-07     NA 0.000 24.184
#> 73  2004-07     NA 0.000 24.196
#> 74  2004-07     NA 0.000 24.305
#> 75  2004-07     NA 0.000 24.280
#> 76  2004-07     NA 0.000 24.341
#> 77  2004-07     NA 0.000 24.290
#> 78  2004-07     NA 0.000 24.567
#> 79  2004-07     NA 0.000 24.657
#> 80  2004-07     NA 0.000 24.523
#> 81  2004-07     NA 0.000 24.327
#> 82  2004-07     NA 9.116 24.185
#> 83  2004-07     NA 0.000 25.381
#> 84  2004-07     NA 0.000 25.151
#> 85  2004-07     NA 0.000 25.011
#> 86  2004-07     NA 0.000 24.714
#> 87  2004-07     NA 0.000 24.880
#> 88  2004-07     NA 0.000 24.794
#> 89  2004-07     NA 0.000 24.574
#> 90  2004-07     NA 0.000 24.633
#> 91  2004-07     NA 0.000 24.170
#> 92  2004-07     NA 0.000 24.213
#> 93  2004-07     NA 0.000 24.135
#> 94  2004-07     NA 0.000 23.986
#> 95  2004-07     NA 0.000 24.050
#> 96  2004-07     NA 0.000 24.139
#> 97  2004-07     NA 0.000 24.300
#> 98  2004-07     NA 0.000 24.390
#> 99  2004-07     NA 0.000 24.303
#> 100 2004-07     NA 0.000 24.613
#> 101 2004-07     NA 0.000 25.087
#> 102 2004-07     NA 0.000 25.143
#> 103 2004-07     NA 0.000 24.964
#> 104 2004-07     NA 0.000 24.623
#> 105 2004-07     NA 7.551 24.403
#> 106 2004-07     NA 0.000 26.660
#> 107 2004-07     NA 0.000 26.085
#> 108 2004-07     NA 0.000 25.754
#> 109 2004-07     NA 0.000 25.520
#> 110 2004-07     NA 0.000 25.372
#> 111 2004-07     NA 0.000 25.318
#> 112 2004-07     NA 0.000 25.396
#> 113 2004-07     NA 0.000 25.120
#> 114 2004-07     NA 0.000 24.904
#> 115 2004-07     NA 0.000 24.678
#> 116 2004-07     NA 0.000 24.657
#> 117 2004-07     NA 0.000 24.510
#> 118 2004-07     NA 0.000 24.287
#> 119 2004-07 10.663 0.000 23.941
#> 120 2004-07     NA 0.000 23.751
#> 121 2004-07     NA 0.000 23.873
#> 122 2004-07     NA 0.000 23.817
#> 123 2004-07     NA 0.000 23.690
#> 124 2004-07     NA 0.000 23.747
#> 125 2004-07     NA 0.000 23.756
#> 126 2004-07     NA 0.000 23.816
#> 127 2004-07     NA 0.000 23.790
#> 128 2004-07     NA 7.200 23.662
#> 129 2004-07  1.129 0.000 29.157
#> 130 2004-07     NA 0.000 28.628
#> 131 2004-07     NA 0.000 28.114
#> 132 2004-07     NA 0.000 27.461
#> 133 2004-07     NA 0.000 27.057
#> 134 2004-07     NA 0.000 26.985
#> 135 2004-07     NA 0.000 26.742
#> 136 2004-07     NA 0.000 26.327
#> 137 2004-07     NA 0.000 25.651
#> 138 2004-07     NA 0.000 25.008
#> 139 2004-07     NA 0.000 24.584
#> 140 2004-07     NA 0.000 24.301
#> 141 2004-07     NA 0.000 23.917
#> 142 2004-07     NA 0.000 23.818
#> 143 2004-07     NA 0.000 23.851
#> 144 2004-07     NA 0.000 23.620
#> 145 2004-07     NA 0.000 23.588
#> 146 2004-07     NA 0.000 23.820
#> 147 2004-07     NA 0.000 23.801
#> 148 2004-07     NA 0.000 23.559
#> 149 2004-07     NA 0.000 23.833
#> 150 2004-07     NA 0.000 23.805
#> 151 2004-07     NA 0.000 28.627
#> 152 2004-07     NA 0.000 27.709
#> 153 2004-07     NA 0.000 26.952
#> 154 2004-07     NA 0.000 26.328
#> 155 2004-07     NA 0.000 25.942
#> 156 2004-07     NA 0.000 25.554
#> 157 2004-07     NA 0.000 25.214
#> 158 2004-07     NA 0.000 25.042
#> 159 2004-07     NA 0.000 24.955
#> 160 2004-07     NA 0.000 24.682
#> 161 2004-07     NA 0.000 24.639
#> 162 2004-07     NA 0.000 25.066
#> 163 2004-07     NA 0.000 24.692
#> 164 2004-07     NA 0.000 24.628
#> 165 2004-07     NA 0.000 24.094
#> 166 2004-07     NA 0.000 23.737
#> 167 2004-07     NA 0.000 23.793
#> 168 2004-07     NA 0.000 23.631
#> 169 2004-07     NA 0.000 23.429
#> 170 2004-07     NA 0.000 23.271
#> 171 2004-07     NA 0.000 23.432
#> 172 2004-07     NA 0.000 23.267
#> 173 2004-07     NA 6.065 23.290
#> 174 2004-07  0.132 0.000 30.131
#> 175 2004-07  1.198 0.000 29.444
#> 176 2004-07  1.342 0.000 28.813
#> 177 2004-07     NA 0.000 28.070
#> 178 2004-07  1.684 0.000 27.722
#> 179 2004-07     NA 0.000 27.323
#> 180 2004-07     NA 0.000 26.680
#> 181 2004-07     NA 0.000 25.828
#> 182 2004-07     NA 0.000 25.344
#> 183 2004-07     NA 0.000 25.466
#> 184 2004-07     NA 0.000 25.467
#> 185 2004-07  1.706 0.000 25.636
#> 186 2004-07     NA 0.000 25.651
#> 187 2004-07     NA 0.000 25.456
#> 188 2004-07     NA 0.000 25.256
#> 189 2004-07     NA 0.000 25.156
#> 190 2004-07     NA 0.000 24.991
#> 191 2004-07     NA 0.000 24.945
#> 192 2004-07     NA 0.000 24.991
#> 193 2004-07     NA 0.000 24.773
#> 194 2004-07     NA 0.000 24.764
#> 195 2004-07     NA 0.000 24.768
#> 196 2004-07     NA 3.804 24.536
#> 197 2004-07     NA 0.000 30.372
#> 198 2004-07     NA 0.000 29.237
#> 199 2004-07     NA 0.000 28.461
#> 200 2004-07     NA 0.000 27.851
#> 201 2004-07  4.593 0.000 27.389
#> 202 2004-07  3.137 0.000 27.106
#> 203 2004-07     NA 0.000 26.928
#> 204 2004-07     NA 0.000 26.686
#> 205 2004-07     NA 0.000 26.236
#> 206 2004-07     NA 0.000 25.954
#> 207 2004-07     NA 0.000 25.732
#> 208 2004-07     NA 0.000 25.441
#> 209 2004-07     NA 0.000 25.248
#> 210 2004-07     NA 0.000 25.204
#> 211 2004-07     NA 0.000 25.128
#> 212 2004-07     NA 0.000 25.098
#> 213 2004-07     NA 0.000 25.219
#> 214 2004-07     NA 0.000 25.269
#> 215 2004-07     NA 0.000 25.216
#> 216 2004-07     NA 0.000 25.195
#> 217 2004-07     NA 0.000 25.123
#> 218 2004-07     NA 0.000 25.100
#> 219 2004-07     NA 6.896 25.165
#> 220 2004-07     NA 0.000 23.723
#> 221 2004-07     NA 0.000 23.623
#> 222 2004-07     NA 0.000 23.670
#> 223 2004-07     NA 0.000 23.948
#> 224 2004-07     NA 0.000 23.790
#> 225 2004-07     NA 0.000 23.490
#> 226 2004-07     NA 0.000 23.280
#> 227 2004-07     NA 0.000 23.050
#> 228 2004-07     NA 0.000 23.016
#> 229 2004-07     NA 0.000 23.151
#> 230 2004-07     NA 0.000 23.232
#> 231 2004-07     NA 0.000 22.996
#> 232 2004-07     NA 0.000 22.886
#> 233 2004-07  8.524 0.000 22.702
#> 234 2004-07  9.591 0.000 22.707
#> 235 2004-07     NA 0.000 22.853
#> 236 2004-07     NA 0.000 22.881
#> 237 2004-07  8.993 0.000 22.655
#> 238 2004-07     NA 0.000 22.610
#> 239 2004-07     NA 0.000 22.731
#> 240 2004-07     NA 0.000 22.799
#> 241 2004-07     NA 0.000 22.699
#> 242 2004-07     NA 5.126 22.592
#> 243 2004-07     NA 6.538 21.487
#> 244 2004-07     NA 4.797 22.264
#> 245 2004-07     NA 5.883 22.088
#> 246 2004-07     NA 0.000 23.585
#> 247 2004-07     NA 0.000 23.334
#> 248 2004-07     NA 0.000 23.132
#> 249 2004-07     NA 0.000 23.043
#> 250 2004-07     NA 0.000 23.073
#> 251 2004-07     NA 0.000 23.131
#> 252 2004-07     NA 0.000 23.018
#> 253 2004-07     NA 0.000 22.911
#> 254 2004-07     NA 0.000 22.655
#> 255 2004-07     NA 0.000 22.432
#> 256 2004-07     NA 0.000 22.230
#> 257 2004-07     NA 0.000 22.227
#> 258 2004-07     NA 0.000 22.195
#> 259 2004-07     NA 0.000 22.041
#> 260 2004-07     NA 0.000 21.921
#> 261 2004-07     NA 0.000 21.953
#> 262 2004-07     NA 0.000 21.990
#> 263 2004-07     NA 0.000 21.869
#> 264 2004-07     NA 0.000 21.817
#> 265 2004-07     NA 0.000 21.780
#> 266 2004-07     NA 0.000 21.613
#> 267 2004-07     NA 0.000 21.417
#> 268 2004-07     NA 4.694 21.612
#> 269 2004-07     NA 0.000 28.319
#> 270 2004-07     NA 0.000 27.443
#> 271 2004-07     NA 0.000 26.724
#> 272 2004-07     NA 0.000 26.582
#> 273 2004-07     NA 0.000 26.146
#> 274 2004-07     NA 0.000 25.672
#> 275 2004-07     NA 0.000 25.925
#> 276 2004-07     NA 0.000 26.034
#> 277 2004-07     NA 0.000 25.434
#> 278 2004-07     NA 0.000 25.333
#> 279 2004-07     NA 0.000 25.069
#> 280 2004-07     NA 0.000 24.851
#> 281 2004-07     NA 0.000 24.513
#> 282 2004-07     NA 0.000 24.458
#> 283 2004-07     NA 0.000 24.221
#> 284 2004-07     NA 0.000 24.322
#> 285 2004-07     NA 0.000 24.575
#> 286 2004-07     NA 0.000 24.362
#> 287 2004-07     NA 0.000 23.858
#> 288 2004-07     NA 0.000 23.496
#> 289 2004-07     NA 0.000 23.355
#> 290 2004-07     NA 0.000 23.255
#> 291 2004-07     NA 1.939 23.331
#> 292 2004-07     NA 0.000 28.586
#> 293 2004-07     NA 0.000 28.184
#> 294 2004-07     NA 0.000 27.464
#> 295 2004-07     NA 0.000 26.852
#> 296 2004-07     NA 0.000 26.577
#> 297 2004-07     NA 0.000 26.491
#> 298 2004-07     NA 0.000 26.460
#> 299 2004-07     NA 0.000 26.105
#> 300 2004-07     NA 0.000 26.645
#> 301 2004-07     NA 0.000 26.561
#> 302 2004-07     NA 0.000 26.350
#> 303 2004-07     NA 0.000 26.271
#> 304 2004-07     NA 0.000 25.732
#> 305 2004-07     NA 0.000 25.328
#> 306 2004-07     NA 0.000 25.363
#> 307 2004-07     NA 0.000 25.247
#> 308 2004-07     NA 0.000 25.181
#> 309 2004-07     NA 0.000 25.098
#> 310 2004-07     NA 0.000 24.964
#> 311 2004-07     NA 0.000 24.900
#> 312 2004-07     NA 0.000 24.870
#> 313 2004-07     NA 0.000 24.880
#> 314 2004-07     NA 2.590 24.735
#> 315 2004-07  1.092 0.000 29.701
#> 316 2004-07  3.176 0.000 28.990
#> 317 2004-07  3.292 0.000 28.403
#> 318 2004-07     NA 0.000 27.972
#> 319 2004-07     NA 0.000 27.639
#> 320 2004-07     NA 0.000 26.966
#> 321 2004-07     NA 0.000 27.353
#> 322 2004-07     NA 0.000 26.682
#> 323 2004-07     NA 0.000 27.239
#> 324 2004-07     NA 0.000 26.812
#> 325 2004-07     NA 0.000 26.618
#> 326 2004-07     NA 0.000 26.342
#> 327 2004-07     NA 0.000 26.651
#> 328 2004-07     NA 0.000 26.397
#> 329 2004-07     NA 0.000 26.434
#> 330 2004-07     NA 0.000 26.261
#> 331 2004-07     NA 0.000 26.233
#> 332 2004-07     NA 0.000 26.124
#> 333 2004-07     NA 0.000 25.673
#> 334 2004-07     NA 0.000 25.930
#> 335 2004-07     NA 0.000 26.242
#> 336 2004-07     NA 0.000 25.431
#> 337 2004-07     NA 3.624 25.233
#> 338 2004-07     NA 0.000 28.750
#> 339 2004-07     NA 0.000 28.776
#> 340 2004-07     NA 0.000 28.451
#> 341 2004-07     NA 0.000 28.190
#> 342 2004-07     NA 0.000 27.950
#> 343 2004-07     NA 0.000 27.830
#> 344 2004-07     NA 0.000 27.100
#> 345 2004-07     NA 0.000 26.757
#> 346 2004-07     NA 0.000 26.407
#> 347 2004-07     NA 0.000 26.161
#> 348 2004-07     NA 0.000 26.519
#> 349 2004-07  0.711 0.000 26.809
#> 350 2004-07     NA 0.000 26.218
#> 351 2004-07     NA 0.000 26.032
#> 352 2004-07     NA 0.000 25.958
#> 353 2004-07     NA 0.000 25.968
#> 354 2004-07     NA 0.000 25.574
#> 355 2004-07     NA 0.000 25.404
#> 356 2004-07     NA 0.000 25.498
#> 357 2004-07     NA 0.000 25.863
#> 358 2004-07     NA 0.000 25.576
#> 359 2004-07     NA 0.000 25.563
#> 360 2004-07     NA 0.000 25.553
#> 361 2004-07     NA 3.506 25.481
#> 362 2004-07     NA 0.000 29.806
#> 363 2004-07     NA 0.000 28.393
#> 364 2004-07     NA 0.000 26.729
#> 365 2004-07     NA 0.000 25.822
#> 366 2004-07     NA 0.000 25.443
#> 367 2004-07     NA 0.000 25.232
#> 368 2004-07  3.487 0.000 24.960
#> 369 2004-07     NA 0.000 24.731
#> 370 2004-07     NA 0.000 24.619
#> 371 2004-07     NA 0.000 24.493
#> 372 2004-07     NA 0.000 24.584
#> 373 2004-07     NA 0.000 24.474
#> 374 2004-07     NA 0.000 24.419
#> 375 2004-07     NA 0.000 24.301
#> 376 2004-07  3.713 0.000 24.367
#> 377 2004-07     NA 0.000 24.312
#> 378 2004-07     NA 0.000 24.372
#> 379 2004-07     NA 0.000 24.728
#> 380 2004-07     NA 0.000 24.846
#> 381 2004-07     NA 0.000 24.817
#> 382 2004-07     NA 0.000 24.939
#> 383 2004-07     NA 0.000 24.880
#> 384 2004-07     NA 0.000 25.337
#> 385 2004-07     NA 2.989 25.120
#> 386 2004-07 -0.802 0.000 29.595
#> 387 2004-07  2.905 0.000 29.004
#> 388 2004-07  2.945 0.000 28.743
#> 389 2004-07     NA 0.000 27.538
#> 390 2004-07     NA 0.000 25.055
#> 391 2004-07     NA 0.000 25.208
#> 392 2004-07     NA 0.000 25.714
#> 393 2004-07     NA 0.000 25.801
#> 394 2004-07     NA 0.000 25.389
#> 395 2004-07     NA 0.000 25.428
#> 396 2004-07     NA 0.000 26.022
#> 397 2004-07     NA 0.000 26.162
#> 398 2004-07     NA 0.000 26.459
#> 399 2004-07  6.515 0.000 26.490
#> 400 2004-07 -1.965 0.000 25.694
#> 401 2004-07     NA 0.000 25.912
#> 402 2004-07     NA 0.000 26.183
#> 403 2004-07  1.854 0.000 26.198
#> 404 2004-07 14.794 0.000 26.211
#> 405 2004-07 10.912 0.000 25.829
#> 406 2004-07 12.455 0.000 25.632
#> 407 2004-07     NA 0.000 25.120
#> 408 2004-07     NA 7.338 23.944
#> 409 2004-07     NA 9.658 23.183
#> 410 2004-07     NA 9.572 23.680
#> 411 2004-07     NA 0.000 23.182
#> 412 2004-07     NA 0.000 23.437
#> 413 2004-07     NA 0.000 23.809
#> 414 2004-07     NA 0.000 24.409
#> 415 2004-07     NA 0.000 24.075
#> 416 2004-07     NA 0.000 23.788
#> 417 2004-07     NA 0.000 23.737
#> 418 2004-07     NA 0.000 23.685
#> 419 2004-07     NA 0.000 24.145
#> 420 2004-07     NA 0.000 24.508
#> 421 2004-07     NA 0.000 24.690
#> 422 2004-07     NA 0.000 25.007
#> 423 2004-07     NA 0.000 25.257
#> 424 2004-07     NA 0.000 25.340
#> 425 2004-07     NA 0.000 25.326
#> 426 2004-07     NA 0.000 25.378
#> 427 2004-07     NA 0.000 25.713
#> 428 2004-07     NA 0.000 25.423
#> 429 2004-07     NA 0.000 25.653
#> 430 2004-07     NA 0.000 25.716
#> 431 2004-07     NA 0.000 25.126
#> 432 2004-07     NA 1.913 24.861
#> 433 2004-07     NA 4.747 24.126
#> 434 2004-07 -1.817 0.000 29.009
#> 435 2004-07     NA 0.000 28.552
#> 436 2004-07     NA 0.000 26.606
#> 437 2004-07     NA 0.000 23.968
#> 438 2004-07     NA 0.000 23.664
#> 439 2004-07     NA 0.000 23.754
#> 440 2004-07     NA 0.000 23.841
#> 441 2004-07     NA 0.000 23.763
#> 442 2004-07     NA 0.000 23.889
#> 443 2004-07     NA 0.000 24.165
#> 444 2004-07     NA 0.000 24.589
#> 445 2004-07     NA 0.000 24.830
#> 446 2004-07     NA 0.000 24.567
#> 447 2004-07     NA 0.000 24.253
#> 448 2004-07     NA 0.000 24.032
#> 449 2004-07     NA 0.000 23.882
#> 450 2004-07     NA 0.000 23.720
#> 451 2004-07     NA 0.000 24.001
#> 452 2004-07     NA 0.000 24.341
#> 453 2004-07     NA 0.000 24.176
#> 454 2004-07     NA 0.000 24.020
#> 455 2004-07     NA 0.000 23.976
#> 456 2004-07     NA 0.000 23.753
#> 457 2004-07     NA 0.184 23.842
#> 458 2004-07     NA 1.483 23.716
#> 459 2004-07     NA 1.959 23.002
#> 460 2004-07     NA 6.546 22.983
#> 461 2004-07     NA 0.000 27.533
#> 462 2004-07  6.402 0.000 26.913
#> 463 2004-07     NA 0.000 26.102
#> 464 2004-07     NA 0.000 25.762
#> 465 2004-07     NA 0.000 25.246
#> 466 2004-07     NA 0.000 25.164
#> 467 2004-07     NA 0.000 25.181
#> 468 2004-07     NA 0.000 25.182
#> 469 2004-07     NA 0.000 25.000
#> 470 2004-07     NA 0.000 24.463
#> 471 2004-07     NA 0.000 23.139
#> 472 2004-07     NA 0.000 22.992
#> 473 2004-07     NA 0.000 22.879
#> 474 2004-07     NA 0.000 22.801
#> 475 2004-07     NA 0.000 22.744
#> 476 2004-07     NA 0.000 22.693
#> 477 2004-07     NA 0.000 22.658
#> 478 2004-07     NA 0.000 22.642
#> 479 2004-07     NA 0.000 22.610
#> 480 2004-07     NA 0.000 22.520
#> 481 2004-07     NA 0.000 22.350
#> 482 2004-07     NA 0.000 22.114
#> 483 2004-07     NA 0.000 21.955
#> 484 2004-07     NA 2.499 22.032
#> 485 2004-07     NA 0.000 23.797
#> 486 2004-07     NA 0.000 23.794
#> 487 2004-07     NA 0.000 23.719
#> 488 2004-07     NA 0.000 23.621
#> 489 2004-07     NA 0.000 23.550
#> 490 2004-07     NA 0.000 23.447
#> 491 2004-07     NA 0.000 23.254
#> 492 2004-07     NA 0.000 23.052
#> 493 2004-07     NA 0.000 22.930
#> 494 2004-07     NA 0.000 25.458
#> 495 2004-07     NA 0.000 25.293
#> 496 2004-07     NA 0.000 25.071
#> 497 2004-07     NA 0.000 24.911
#> 498 2004-07     NA 0.000 24.652
#> 499 2004-07     NA 0.000 24.549
#> 500 2004-07     NA 0.000 24.391
#> 501 2004-07     NA 0.000 24.277
#> 502 2004-07     NA 0.000 24.193
#> 503 2004-07     NA 0.000 24.370
#> 504 2004-07     NA 0.000 24.212
#> 505 2004-07     NA 0.000 24.301
#> 506 2004-07     NA 0.000 24.379
#> 507 2004-07     NA 3.097 24.442
#> 508 2004-07  1.682 0.000 27.178
#> 509 2004-07  4.115 0.000 26.792
#> 510 2004-07  3.676 0.000 26.602
#> 511 2004-07  3.854 0.000 26.683
#> 512 2004-07     NA 0.000 26.643
#> 513 2004-07     NA 0.000 26.432
#> 514 2004-07     NA 0.000 26.097
#> 515 2004-07     NA 0.000 25.863
#> 516 2004-07     NA 0.000 25.410
#> 517 2004-07     NA 0.000 25.103
#> 518 2004-07     NA 0.000 24.899
#> 519 2004-07     NA 0.000 24.647
#> 520 2004-07     NA 0.000 24.563
#> 521 2004-07     NA 0.000 24.539
#> 522 2004-07     NA 0.000 24.356
#> 523 2004-07     NA 0.000 24.386
#> 524 2004-07     NA 0.000 24.276
#> 525 2004-07     NA 0.000 24.177
#> 526 2004-07     NA 0.000 24.133
#> 527 2004-07     NA 0.000 23.882
#> 528 2004-07     NA 0.000 23.746
#> 529 2004-07     NA 0.000 23.745
#> 530 2004-07     NA 0.000 23.844
#> 531 2004-07     NA 3.091 24.031
#> 532 2004-07     NA 0.000 28.704
#> 533 2004-07  2.686 0.000 28.359
#> 534 2004-07  2.035 0.000 27.932
#> 535 2004-07     NA 0.000 27.544
#> 536 2004-07  3.118 0.000 27.094
#> 537 2004-07     NA 0.000 26.709
#> 538 2004-07     NA 0.000 26.314
#> 539 2004-07     NA 0.000 26.066
#> 540 2004-07     NA 0.000 25.803
#> 541 2004-07     NA 0.000 25.562
#> 542 2004-07     NA 0.000 25.340
#> 543 2004-07     NA 0.000 25.069
#> 544 2004-07     NA 0.000 24.926
#> 545 2004-07     NA 0.000 24.839
#> 546 2004-07     NA 0.000 24.866
#> 547 2004-07     NA 0.000 24.458
#> 548 2004-07     NA 0.000 24.152
#> 549 2004-07     NA 0.000 24.107
#> 550 2004-07     NA 0.000 24.217
#> 551 2004-07     NA 0.000 24.170
#> 552 2004-07     NA 0.000 23.893
#> 553 2004-07     NA 0.000 23.866
#> 554 2004-07     NA 1.586 23.875
#> 555 2004-07     NA 0.000 23.438
#> 556 2004-07     NA 0.000 23.251
#> 557 2004-07     NA 0.000 23.533
#> 558 2004-07     NA 0.000 23.631
#> 559 2004-07     NA 0.000 23.801
#> 560 2004-07     NA 0.000 23.964
#> 561 2004-07     NA 0.000 24.284
#> 562 2004-07     NA 0.000 24.290
#> 563 2004-07     NA 0.000 24.088
#> 564 2004-07     NA 0.000 24.074
#> 565 2004-07     NA 0.000 24.075
#> 566 2004-07     NA 0.000 24.028
#> 567 2004-07     NA 0.000 24.178
#> 568 2004-07     NA 0.000 24.216
#> 569 2004-07     NA 0.000 24.095
#> 570 2004-07     NA 0.000 24.053
#> 571 2004-07     NA 0.000 24.011
#> 572 2004-07     NA 0.000 23.855
#> 573 2004-07     NA 0.000 23.745
#> 574 2004-07     NA 0.000 23.727
#> 575 2004-07     NA 0.000 23.656
#> 576 2004-07     NA 0.000 23.807
#> 577 2004-07     NA 0.000 23.923
#> 578 2004-07     NA 0.000 27.051
#> 579 2004-07     NA 0.000 26.788
#> 580 2004-07     NA 0.000 26.203
#> 581 2004-07     NA 0.000 25.552
#> 582 2004-07     NA 0.000 25.295
#> 583 2004-07     NA 0.000 25.371
#> 584 2004-07     NA 0.000 25.410
#> 585 2004-07     NA 0.000 25.024
#> 586 2004-07     NA 0.000 24.552
#> 587 2004-07     NA 0.000 24.368
#> 588 2004-07     NA 0.000 24.255
#> 589 2004-07     NA 0.000 24.239
#> 590 2004-07     NA 0.000 24.389
#> 591 2004-07     NA 0.000 24.303
#> 592 2004-07     NA 0.000 23.967
#> 593 2004-07     NA 0.000 24.063
#> 594 2004-07     NA 0.000 24.293
#> 595 2004-07     NA 0.000 24.485
#> 596 2004-07     NA 0.000 24.391
#> 597 2004-07     NA 0.000 24.044
#> 598 2004-07     NA 0.000 23.829
#> 599 2004-07     NA 0.000 23.418
#> 600 2004-07     NA 0.000 23.207
#> 601 2004-07     NA 2.477 23.371
#> 602 2004-07     NA 9.745 26.901
#> 603 2004-07     NA 5.686 24.599
#> 604 2004-07  7.263 8.519 23.676
#> 605 2004-07     NA 0.000 23.451
#> 606 2004-07     NA 0.000 23.479
#> 607 2004-07     NA 0.000 23.286
#> 608 2004-07     NA 0.000 23.212
#> 609 2004-07     NA 0.000 23.064
#> 610 2004-07     NA 0.000 23.051
#> 611 2004-07     NA 0.000 23.129
#> 612 2004-07     NA 0.000 23.156
#> 613 2004-07     NA 0.000 23.415
#> 614 2004-07     NA 0.000 23.442
#> 615 2004-07     NA 0.000 23.101
#> 616 2004-07     NA 0.000 23.014
#> 617 2004-07     NA 0.000 23.189
#> 618 2004-07     NA 0.000 23.156
#> 619 2004-07     NA 0.000 23.329
#> 620 2004-07     NA 0.000 23.306
#> 621 2004-07     NA 0.000 23.289
#> 622 2004-07     NA 0.000 23.276
#> 623 2004-07     NA 0.000 23.128
#> 624 2004-07     NA 0.000 23.162
#> 625 2004-07     NA 0.000 23.192
#> 626 2004-07     NA 0.000 23.243
#> 627 2004-07     NA 0.000 23.271
#> 628 2004-07     NA 0.591 23.360
#> 629 2004-07     NA 0.000 27.154
#> 630 2004-07     NA 0.000 26.015
#> 631 2004-07     NA 0.000 25.835
#> 632 2004-07     NA 0.000 25.657
#> 633 2004-07     NA 0.000 25.431
#> 634 2004-07     NA 0.000 24.487
#> 635 2004-07     NA 0.000 24.173
#> 636 2004-07     NA 0.000 24.060
#> 637 2004-07     NA 0.000 24.044
#> 638 2004-07     NA 0.000 24.113
#> 639 2004-07     NA 0.000 23.962
#> 640 2004-07     NA 0.000 23.758
#> 641 2004-07     NA 0.000 23.632
#> 642 2004-07     NA 0.000 23.667
#> 643 2004-07     NA 0.000 23.780
#> 644 2004-07     NA 0.000 23.718
#> 645 2004-07     NA 0.000 23.695
#> 646 2004-07     NA 0.000 23.656
#> 647 2004-07     NA 0.000 23.649
#> 648 2004-07     NA 0.000 24.033
#> 649 2004-07     NA 0.000 24.055
#> 650 2004-07     NA 0.000 23.794
#> 651 2004-07     NA 0.000 23.573
#> 652 2004-07     NA 0.499 23.509
#> 653 2004-07     NA 0.000 26.163
#> 654 2004-07     NA 0.000 25.947
#> 655 2004-07     NA 0.000 25.537
#> 656 2004-07     NA 0.000 25.335
#> 657 2004-07     NA 0.000 25.119
#> 658 2004-07     NA 0.000 24.699
#> 659 2004-07     NA 0.000 24.899
#> 660 2004-07     NA 0.000 24.715
#> 661 2004-07     NA 0.000 24.904
#> 662 2004-07     NA 0.000 24.755
#> 663 2004-07     NA 0.000 24.679
#> 664 2004-07     NA 0.000 24.596
#> 665 2004-07     NA 0.000 24.724
#> 666 2004-07     NA 0.000 24.664
#> 667 2004-07     NA 0.000 24.708
#> 668 2004-07     NA 0.000 24.527
#> 669 2004-07     NA 0.000 24.511
#> 670 2004-07     NA 0.000 24.293
#> 671 2004-07     NA 0.000 24.028
#> 672 2004-07     NA 0.000 23.948
#> 673 2004-07     NA 0.000 23.954
#> 674 2004-07     NA 0.000 23.859
#> 675 2004-07     NA 0.000 23.753
#> 676 2004-07     NA 0.754 23.553
#> 677 2004-07     NA 0.000 27.002
#> 678 2004-07  1.259 0.000 26.657
#> 679 2004-07  2.217 0.000 26.447
#> 680 2004-07     NA 0.000 26.221
#> 681 2004-07     NA 0.000 26.014
#> 682 2004-07     NA 0.000 25.692
#> 683 2004-07     NA 0.000 25.612
#> 684 2004-07     NA 0.000 25.530
#> 685 2004-07     NA 0.000 25.420
#> 686 2004-07     NA 0.000 25.320
#> 687 2004-07     NA 0.000 25.220
#> 688 2004-07     NA 0.000 24.943
#> 689 2004-07     NA 0.000 24.647
#> 690 2004-07     NA 0.000 24.521
#> 691 2004-07     NA 0.000 24.488
#> 692 2004-07     NA 0.000 24.585
#> 693 2004-07     NA 0.000 24.493
#> 694 2004-07     NA 0.000 24.453
#> 695 2004-07     NA 0.000 24.250
#> 696 2004-07     NA 0.000 24.167
#> 697 2004-07     NA 0.000 23.997
#> 698 2004-07     NA 0.000 23.863
#> 699 2004-07     NA 0.000 23.735
#> 700 2004-07     NA 0.944 23.770
#> 701 2004-07     NA 9.373 23.926
#> 702 2004-07  3.315 0.000 28.161
#> 703 2004-07  1.571 0.000 27.866
#> 704 2004-07  3.981 0.000 27.670
#> 705 2004-07  3.346 0.000 27.486
#> 706 2004-07     NA 0.000 27.273
#> 707 2004-07     NA 0.000 26.858
#> 708 2004-07     NA 0.000 26.557
#> 709 2004-07     NA 0.000 25.915
#> 710 2004-07     NA 0.000 25.766
#> 711 2004-07     NA 0.000 25.560
#> 712 2004-07  5.595 0.000 25.388
#> 713 2004-07     NA 0.000 25.064
#> 714 2004-07     NA 0.000 24.859
#> 715 2004-07     NA 0.000 24.895
#> 716 2004-07     NA 0.000 24.675
#> 717 2004-07     NA 0.000 24.819
#> 718 2004-07     NA 0.000 24.753
#> 719 2004-07     NA 0.000 24.618
#> 720 2004-07     NA 0.000 24.522
#> 721 2004-07     NA 0.000 24.507
#> 722 2004-07     NA 0.000 24.754
#> 723 2004-07     NA 0.000 25.373
#> 724 2004-07     NA 0.313 25.555
#> 725 2004-07     NA 0.000 26.307
#> 726 2004-07     NA 0.000 26.092
#> 727 2004-07 11.472 0.000 25.429
#> 728 2004-07  7.672 0.000 23.551
#> 729 2004-07     NA 0.000 23.446
#> 730 2004-07     NA 0.000 23.554
#> 731 2004-07     NA 0.000 24.405
#> 732 2004-07     NA 0.000 24.240
#> 733 2004-07     NA 0.000 24.387
#> 734 2004-07     NA 0.000 24.496
#> 735 2004-07     NA 0.000 24.413
#> 736 2004-07     NA 0.000 24.316
#> Warning: Rows containing NAs were excluded from the model.
#> Error : CmdStan path has not been set yet. See ?set_cmdstan_path.
#> Error in TRC_PARMS_06(data.frame = tower.data, iterations = 5000, priors.trc = brms::prior("normal(0.5, 0.3)",     nlpar = "Ea", lb = 0.01, ub = 1) + brms::prior("normal(.5, .3)",     nlpar = "Rref", lb = 0.01, ub = 1), idx.colname = "YearMon",     NEE.colname = "NEE_PI", PAR.colname = "SW_IN", TA.colname = "TA_1_1_1"): object 'model.brms' not found